Added SRS Doc
Feel free to add or modify. As I started working on it, it occurred to
me that the system description (4.1) and the user documentation (5.4)
were going to be quite similar. I'm not sure if it should be this way,
so if you guys have anything to add, that would be great. Thanks!